Refactor WebsocketEndpoint and extract WebRTCGStreamerPlugInEventsHandler methods
This commit separates the functional logic in the WebsocketEndpoint by extracting the event handling methods such as resume, decreaseVolume, increaseVolume, and pause into their own dedicated methods in the new WebRTCGStreamerPlugInEventsHandler class. The event handling now verifies if the feature equals "G_STREAM_BROADCAST". This results in a cleaner code structure, promotes method reusability and improves the readability of the code. It also simplifies the process of maintaining and enhancing the event handling functionality in the future.